Added python and pickle files
The python file is what I wrote to create final figures for this project showing foldchange in selected genes. The pickled dictionary is a transcript-to-genename dictionary that was necessary to use while labeling genes with adjusted p-values, because the p-value data only showed transcript IDs rather than gene names (multiple transcripts can belong tot he same gene, so transcripts are most important to keep track of during a transcriptomics project).